Crate piet_embedded_graphics

The embedded-graphics backend for the Piet 2D graphics abstraction.

Modules

batch
Batch the pixels to be rendered into Pixel Rows and Pixel Blocks (contiguous Pixel Rows). This enables the pixels to be rendered efficiently as Pixel Blocks, which may be transmitted in a single Non-Blocking SPI request.

brush

context

display

image

status

text

Structs

EmbedFont

EmbedFontBuilder

EmbedRenderContext

EmbedText
Right now, we don't need any state, as the "toy text API" treats the access to system font information as a global. This will change.

EmbedTextLayout

EmbedTextLayoutBuilder

Enums

Brush

Functions

draw_to_display
Draw the item to the display

set_display_pixels
Draw the pixel colours to the display at the window from (sx,sy) to (ex,ey)

show_touch
Display the touched (X, Y) coordinates

start_display
Initialise the display controller