Module piet_embedded_graphics::batch

Batch the pixels to be rendered into Pixel Rows and Pixel Blocks (contiguous Pixel Rows). This enables the pixels to be rendered efficiently as Pixel Blocks, which may be transmitted in a single Non-Blocking SPI request.

Structs

BlockIterator
Iterator for each Pixel Block in the pixel data. A Pixel Block consists of contiguous Pixel Rows with the same start and end column number.

PixelBlock
A block of contiguous pixel rows with the same start and end column number

PixelRow
A row of contiguous pixels

RowIterator
Iterator for each Pixel Row in the pixel data. A Pixel Row consists of contiguous pixels on the same row.

Functions

draw_blocks
Draw the pixels in the item as Pixel Blocks of contiguous Pixel Rows. The pixels are grouped by row then by block.

to_blocks
Batch the Pixel Rows into Pixel Blocks, which are contiguous Pixel Rows with the same start and end column number R can be any Pixel Row Iterator.

to_rows
Batch the pixels into Pixel Rows, which are contiguous pixels on the same row. P can be any Pixel Iterator (e.g. a rectangle).

Type Definitions

BlockColors
Consecutive color words for a Pixel Block

MaxBlockSize
Max number of pixels per Pixel Block

MaxRowSize
Max number of pixels per Pixel Row

RowColors
Consecutive color words for a Pixel Row