Function piet_embedded_graphics::batch::to_blocks

Important traits for BlockIterator<R>
impl<R: Iterator<Item = PixelRow>> Iterator for BlockIterator<R> type Item = PixelBlock;

fn to_blocks<R>(rows: R) -> BlockIterator<R> where
    R: Iterator<Item = PixelRow>, 

Batch the Pixel Rows into Pixel Blocks, which are contiguous Pixel Rows with the same start and end column number R can be any Pixel Row Iterator.